MPI build options and MPI testing.
