+++ meson_options.txt
@@ -1,0 +1,1 @@
+option('USE_MPI', type: 'string', value: 'auto', description: 'Choose MPI implementation (mpich, openmpi, none, auto)')
